refactor connect pool components
