+++ @xen-orchestra/web-core/lib/locales/en.json
@@ -270,7 +270,7 @@
   "job.vm-force-reboot.bad-power-state": "VM must be running or paused",
   "job.vm-force-reboot.in-progress": "Force reboot in progress…",
   "job.vm-force-reboot.missing-vms": "No VM to force reboot",
-  "job.vm-force-shutdown.bad-power-state": "VM must be running, suspended or paused",
+  "job.vm-force-shutdown.bad-power-state": "VM must be running, suspended, or paused",
   "job.vm-force-shutdown.in-progress": "Force shutdown in progress…",
   "job.vm-force-shutdown.missing-vms": "No VM to force shutdown",
   "job.vm-migrate.bad-power-state": "VM must be running",
@@ -395,7 +395,7 @@
   "on": "On",
   "on-object": "on {object}",
   "open-console-in-new-tab": "Open console in new tab",
-  "option": "Option",
+  "options": "Options",
   "or": "Or",
   "os-kernel": "OS kernel",
   "os-name": "OS name",
