+++ @xen-orchestra/web-core/lib/locales/fr.json
@@ -395,7 +395,7 @@
   "on": "Allumé",
   "on-object": "sur {object}",
   "open-console-in-new-tab": "Ouvrir la console dans un nouvel onglet",
-  "option": "Option",
+  "options": "Options",
   "or": "Ou",
   "os-kernel": "Noyau de l'OS",
   "os-name": "Nom de l'OS",
@@ -584,7 +584,7 @@
   "virtualization-boot-settings": "Virtualisation & paramètres de démarrage",
   "virtualization-mode": "Mode de virtualisation",
   "virtualization-type": "Type de virtualisation",
-  "visit-pool-dashboard": "Visiter la page du pool",
+  "visit-pool-dashboard": "Visiter le tableau de bord du pool",
   "vlan": "VLAN",
   "vm": "VM",
   "vm-description": "Description de la VM",
