+++ @xen-orchestra/web/src/pages/pool/connect/index.vue
@@ -1,9 +1,18 @@
 <template>
   <ConnectPoolHeader />
-  <ConnectionCard />
+  <UiCard class="pool-connect-card">
+    <ConnectionForm />
+  </UiCard>
 </template>
 
 <script setup lang="ts">
-import ConnectionCard from '@/components/pool/connect/ConnectionCard.vue'
+import ConnectionForm from '@/components/pool/connect/ConnectionForm.vue'
 import ConnectPoolHeader from '@/components/pool/connect/ConnectPoolHeader.vue'
+import UiCard from '@core/components/ui/card/UiCard.vue'
 </script>
+
+<style lang="postcss" scoped>
+.pool-connect-card {
+  margin: 0.8rem;
+}
+</style>
